sistema_gestao/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
reportlab==4.4.10
Pillow==12.1.1
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
charset-normalizer==3.4.4
